+++ src/components/ui/progress.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as ProgressPrimitive from "@radix-ui/react-progress";
 import Image from "next/image";
-import characterIcon from "../../../public/yellow-face.svg";
+import characterIcon from "../../../public/img/yellow-face.svg";
 
 import { cn } from "@/lib/utils";
